Bugfix: schedule.NextSchedule avoids short jumps
Depending on the parameters, NextSchedule() might have declared
the next scheduled run to be "in a few minutes".

Example: if the previous call subtracted a large jitter from the
target time, and the job finished fast, the target time might not
have been reached yet. In this case the scheduler might chose a
time at the next target time (plus jitter), which then might not
be ~24 hours from now, but only a few minutes in the future.