ssh: remove host name from control socket path
This fixes an issue with long host names, as Unix socket names
must be at most ~100 characters.

By removing the host name from the path, we loose the ability
to distinguish the socket files (which was useful for cleaning
up during development), but gain the ability to backup hosts
with long DNS names.

See: https://unix.stackexchange.com/q/367008